# Enh speed up init emb conv2d