--- a/tests/test_custom_models.py
+++ b/tests/test_custom_models.py
@@ -368,12 +368,44 @@ class PeftCustomModelTester(unittest.TestCase, PeftCommonTester):
 class TestRepr(unittest.TestCase):
     """Tests related to the repr of adapted models"""
 
-    def test_repr_lora(self):
+    def test_repr_lora_linear(self):
         config = LoraConfig(target_modules=["lin0"])
         model = get_peft_model(MLP(), config)
         print_output = repr(model.model.lin0)
         self.assertTrue(print_output.startswith("Linear"))
         self.assertTrue("in_features=10, out_features=20" in print_output)
+        self.assertTrue("lora_A" in print_output)
+        self.assertTrue("lora_B" in print_output)
+        self.assertTrue("default" in print_output)
+
+    def test_repr_lora_embedding(self):
+        config = LoraConfig(target_modules=["emb"])
+        model = get_peft_model(ModelEmbConv1D(), config)
+        print_output = repr(model.model.emb)
+        self.assertTrue(print_output.startswith("Embedding"))
+        self.assertTrue("100, 5" in print_output)
+        self.assertTrue("lora_embedding_A" in print_output)
+        self.assertTrue("lora_embedding_B" in print_output)
+        self.assertTrue("default" in print_output)
+
+    def test_repr_lora_conv1d(self):
+        config = LoraConfig(target_modules=["conv1d"])
+        model = get_peft_model(ModelEmbConv1D(), config)
+        print_output = repr(model.model.conv1d)
+        self.assertTrue(print_output.startswith("Linear"))
+        self.assertTrue("in_features=5, out_features=1" in print_output)
+        self.assertTrue("lora_A" in print_output)
+        self.assertTrue("lora_B" in print_output)
+        self.assertTrue("default" in print_output)
+
+    def test_repr_lora_conv2d(self):
+        config = LoraConfig(target_modules=["conv2d"])
+        model = get_peft_model(ModelConv2D(), config)
+        print_output = repr(model.model.conv2d)
+        self.assertTrue(print_output.startswith("Conv2d"))
+        self.assertTrue("5, 10" in print_output)
+        self.assertTrue("kernel_size=(3, 3)" in print_output)
+        self.assertTrue("stride=(1, 1)" in print_output)
         self.assertTrue("lora_A" in print_output)
         self.assertTrue("lora_B" in print_output)
         self.assertTrue("default" in print_output)