Resolves part of #872

## Description

After getting faster initialization of the LoRA `Linear` layer, initialization of `Conv2d` and `Embedding` is now sped up.

## Implementation

The approach of how to achieve the speed up has slightly changed compared to last time. To refresh memory, in #887, we avoided the unnecessary initialization of the full weight matrix by completely skipping `nn.Linear.__init__`.

Although it is possible to do the same for `Embedding` and `Conv2d`, we run into some trouble here. The issue is that the `__init__` methods of these classes have quite a lot more arguments and some custom logic (i.e. not only `self.foo = foo` but more on top). If we wanted to skip `__init__` entirely, we would have to basically copy all of that into our code. Although that is possible, it is brittle (e.g. the logic could be different for different PyTorch versions or change over time).

For that reason, I opted to implement this differently, using a suggestion we had discussed earlier. The approach is to call `__init__` of the parent class but enforce empty weights (this is what `torch.nn.utils.skip_init` does, although we cannot use that function directly). This way, we can avoid having to copy the `__init__` code while still avoiding expensive initialization of the weights.

I did not change the code for `Linear` to also use this approach because the logic inside of `Linear.__init__` is quite simple (at least for now), so we are good here with the existing approach.

## Benchmark

I was curious how changing the approach for `Linear` would affect the initialization speed. Therefore, I ran the script from #872, 3 times each, with the current vs the new approach.

Current approach:

```
test 1 with model bert-base took 0.021 sec.
test 1 with model bert-base took 0.020 sec.
test 1 with model bert-base took 0.020 sec.
test 2 with model bloomz-1b7 took 0.030 sec.
test 2 with model bloomz-1b7 took 0.030 sec.
test 2 with model bloomz-1b7 took 0.030 sec.
```

New approach if applied to Linear:

```
test 1 with model bert-base took 0.038 sec.
test 1 with model bert-base took 0.039 sec.
test 1 with model bert-base took 0.038 sec.
test 2 with model bloomz-1b7 took 0.072 sec.
test 2 with model bloomz-1b7 took 0.048 sec.
test 2 with model bloomz-1b7 took 0.048 sec.
```

This shows that the new approach is indeed a bit slower than the existing one, though still a lot faster than what we had before. IMHO, I think we're safe to leave the code inside of `Linear` as is and benefit from the slightly better performance at the cost of slightly more fragile code. But please let me know if you prefer:

1. The new approach should also be applied to `Linear`
2. The existing approach should also be applied to `Embedding` and `Conv2d`